tools/config_management/requirements.txt
pydantic>=2.0.0
jsonschema>=4.17.0
PyYAML>=6.0
Jinja2>=3.1.0
toml>=0.10.2
deepdiff>=6.0.0
click>=8.0.0
hvac>=1.0.0
boto3>=1.26.0
gitpython>=3.1.0
pytest>=7.0.0
pytest-cov>=4.0.0
